+++ CHANGES.md
@@ -5,8 +5,10 @@
 - [NEW] Support multiple plugins. See 'api-migration.md' for migration details.
 - [NEW] Allow custom service name in CloudFoundry VCAP_SERVICES environment
   variable.
+- [FIXED] Fix `get_security`/`set_security` asymmetry.
 - [IMPROVED] Updated documentation by replacing deprecated Cloudant links with
   the latest bluemix.net links.
+- [REMOVED] Remove previously deprecated method `set_permissions`.
 
 # 1.10.0 (2017-11-01)
 - [UPGRADED] Upgrade package: cloudant-nano@6.7.0.